Move tests that check public APIs to test directory
