+++ precis-profiles/src/passwords.rs
@@ -122,108 +122,3 @@ impl PrecisFastInvocation for OpaqueString {
         get_opaque_string_profile().compare(s1, s2)
     }
 }
-
-#[cfg(test)]
-mod opaque_string {
-    use crate::passwords::*;
-
-    #[test]
-    fn prepare() {
-        let profile = OpaqueString::new();
-
-        // SPACE `U+0020` is allowed
-        let res = profile.prepare("correct horse battery staple");
-        assert_eq!(res.is_ok(), true);
-        assert_eq!(res.unwrap(), "correct horse battery staple");
-
-        // Differs by case from previous example
-        let res = profile.prepare("Correct Horse Battery Staple");
-        assert_eq!(res.is_ok(), true);
-        assert_eq!(res.unwrap(), "Correct Horse Battery Staple");
-
-        // Non-ASCII letters are OK (e.g., GREEK SMALL LETTER
-        // PI `U+03C0`)
-        let res = profile.prepare("πßå");
-        assert_eq!(res.is_ok(), true);
-        assert_eq!(res.unwrap(), "πßå");
-
-        // Symbols are OK (e.g., BLACK DIAMOND SUIT `U+2666`)
-        let res = profile.prepare("Jack of ♦s");
-        assert_eq!(res.is_ok(), true);
-        assert_eq!(res.unwrap(), "Jack of ♦s");
-
-        // Zero-length passwords are disallowed
-        let res = profile.prepare("");
-        assert_eq!(res.is_err(), true);
-
-        // Control characters like TAB `U+0009` are disallowed
-        let res = profile.prepare("simple;\u{0009} test");
-        assert_eq!(res.is_err(), true);
-    }
-
-    #[test]
-    fn enforce() {
-        let profile = OpaqueString::new();
-
-        // SPACE `U+0020` is allowed
-        let res = profile.enforce("correct horse battery staple");
-        assert_eq!(res.is_ok(), true);
-        assert_eq!(res.unwrap(), "correct horse battery staple");
-
-        // Differs by case from previous example
-        let res = profile.enforce("Correct Horse Battery Staple");
-        assert_eq!(res.is_ok(), true);
-        assert_eq!(res.unwrap(), "Correct Horse Battery Staple");
-
-        // Non-ASCII letters are OK (e.g., GREEK SMALL LETTER
-        // PI `U+03C0`)
-        let res = profile.enforce("πßå");
-        assert_eq!(res.is_ok(), true);
-        assert_eq!(res.unwrap(), "πßå");
-
-        // Symbols are OK (e.g., BLACK DIAMOND SUIT `U+2666`)
-        let res = profile.enforce("Jack of ♦s");
-        assert_eq!(res.is_ok(), true);
-        assert_eq!(res.unwrap(), "Jack of ♦s");
-
-        // `OGHAM` SPACE MARK `U+1680` is mapped to SPACE `U+0020`;
-        // thus, the full string is mapped to <foo bar>
-        let res = profile.enforce("foo bar");
-        assert_eq!(res.is_ok(), true);
-        assert_eq!(res.unwrap(), "foo bar");
-
-        // Symbols are OK (e.g., BLACK DIAMOND SUIT `U+2666`)
-        let res = profile.enforce("Jack of ♦s");
-        assert_eq!(res.is_ok(), true);
-        assert_eq!(res.unwrap(), "Jack of ♦s");
-
-        // Zero-length passwords are disallowed
-        let res = profile.enforce("");
-        assert_eq!(res.is_err(), true);
-
-        // Control characters like TAB `U+0009` are disallowed
-        let res = profile.enforce("simple;\u{0009} test");
-        assert_eq!(res.is_err(), true);
-    }
-
-    #[test]
-    fn compare() {
-        let profile = OpaqueString::new();
-
-        let res = profile.compare("𝄞💝♦💣東💯 Secret", "");
-        assert_eq!(res.is_err(), true);
-
-        let res = profile.compare("", "𝄞💝♦💣東💯 Secret");
-        assert_eq!(res.is_err(), true);
-
-        // Same string. `OGHAM` SPACE MARK `U+1680` is mapped to SPACE `U+0020`
-        let res = profile.compare("𝄞💝♦💣東💯 Secret", "𝄞💝♦💣東💯 Secret");
-        assert_eq!(res.is_ok(), true);
-        assert_eq!(res.unwrap(), true);
-
-        // Differs by case
-        let res = profile.compare("Secret", "secret");
-        assert_eq!(res.is_ok(), true);
-        assert_eq!(res.unwrap(), false);
-    }
-}
